refactor: move feeds from utils to services
