+++ src/pages/[...page]/feed.xml.ts
@@ -8,13 +8,13 @@ import {
   type ListingPageCollection,
   type RelatedEntries,
 } from "../../services/collections";
-import { useI18n } from "../../services/i18n";
-import { CONFIG } from "../../utils/constants";
-import { MissingSiteConfigError } from "../../utils/exceptions";
 import {
   getFeedLanguageFromLocale,
   getRSSItemsFromEntries,
-} from "../../utils/feeds";
+} from "../../services/feeds";
+import { useI18n } from "../../services/i18n";
+import { CONFIG } from "../../utils/constants";
+import { MissingSiteConfigError } from "../../utils/exceptions";
 
 export const getStaticPaths = (async () => {
   const { entries } = await queryCollection(
