+++ src/services/feeds/index.ts
@@ -1,0 +1,2 @@
+export * from "./build-feeds";
+export * from "./collections-feeds";
